requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-socks==0.8.0
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
capmonster-python==2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cli-helpers==2.3.0
click==8.1.3
colorama==0.4.6
colored==1.4.4
Colr==0.9.1
commonmark==0.9.1
configobj==5.0.8
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.0
delegator.py==0.1.1
discord-protos==0.0.2
discord.py==2.3.2
discord.py-self==2.0.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
fonttools==4.50.0
frida==16.1.3
frida-tools==12.1.3
frozenlist==1.3.3
gevent==23.7.0
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
linkify-it-py==2.0.3
litecli==1.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.0.33
names==0.3.0
nextcord==2.5.0
numpy==1.25.1
objection==1.11.0
packaging==24.0
pandas==2.2.1
pexpect==4.8.0
pillow==10.2.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
py-cord==2.4.1
pyarmor==8.2.9
pyarmor.cli.core==3.2.9
pycares==4.3.0
pycparser==2.21
pycryptodome==3.18.0
PyExecJS==1.5.1
Pygments==2.15.1
pyparsing==3.1.2
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
PyTermGUI==7.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-socks==2.3.0
pytweening==1.0.7
pytz==2024.1
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==12.6.0
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==2.13.0
shiboken6==6.6.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
sqlparse==0.4.4
tabulate==0.9.0
terminut==0.0.0.901
textual==0.1.18
threadpoolctl==3.3.0
tls-client==0.2.1
toml==0.10.2
typing_extensions==4.6.3
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.0.3
uvloop==0.17.0
veilcord==0.0.2
wcwidth==0.2.6
websocket==0.2.1
websocket-client==1.6.1
Werkzeug==2.3.6
word2number==1.1
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
